feat: hook summary
